requirements.txt
python-telegram-bot>=22.0
google-generativeai>=0.8.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
